updated variable in lerpcolor from float to double

+++ includes/ft_exdlib.h
@@ -6,7 +6,7 @@
 /*   By: sakitaha <sakitaha@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/25 21:48:55 by sakitaha          #+#    #+#             */
-/*   Updated: 2024/02/19 22:26:23 by sakitaha         ###   ########.fr       */
+/*   Updated: 2024/03/14 13:30:11 by sakitaha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,6 @@
 # include <stdbool.h>
 
 char	*ft_itoa(int n);
-int		ft_lerpcolor(int color_1, int color_2, float t);
+int		ft_lerpcolor(int color_1, int color_2, double t);
 
 #endif
